pyproject.toml
[tool.poetry]
name = "QuizAPI"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.1"}
pydantic = ">=2.7,<3.0"
langchain-openai = "^0.3.3"
python-decouple = "^3.8"
langchain-core = "^0.3.34"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
